Move multihash code out of __init__.py, export public fns & constants
